Bring in Apache Commons File Upload to handle multipart messages
With the addition of this library it is no longer necessary to manually parse the message body. Unfortunately the library usage is limited to one file and could be written directly into the apollo code to clean this up and remove the dependency. However for the sake of keeping things fairly simple and to keep this examples changes isolated that work can remain as something to do in the future.